Refactor: tools: create template for converting crm_report to C
crm_report was ridiculously large and complex for a shell script, horribly
inefficient, and bug-prone. Converting it to C will make it easier to maintain
going forward. Python was considered, but currently Python is only a dependency
of the cts package, and there's the possibility that Python isn't installed on
all cluster nodes.