Refactor: tools: store crm_report from/to times as time_t
Only valid ISO 8601 specifications will be accepted. This is more restrictive
than the previous shell script version of crm_report, which used Perl's
Date::Time module ParseDateString() function, but it accepts the format
specified in the command help.

This also allows us to easily sanity-check that the 'to' time is not earlier
than the 'from' time, which the shell script implementation did not do.