Low: scheduler: consider only optional actions for skipping mandatory ordering
If the 'first' action in a mandatory asymmetric ordering is unrunnable, the
'then' action is normally made required and unrunnable.

This has an exception when the 'then' action would not change the state of its
resource (a stop action for a resource that's already stopped, or a start
action for a resource that's already active).

Previously, this exception would consider only optional starts, since mandatory
starts indicate a restart.

Now, any action (even stops) will be considered for this exception only if
optional.