Refactor: tools: implement crm_report directory management in C
In the previous shell script implementation, the first step crm_report's report
collector would do is search the local host for the location of various
directories (such as the runtime data directory, the location of daemon
executables, etc.). This was necessary because the collector was a shell script
copied from the report initiator, which could have a different directory
layout.

Now that the report collector is an option to crm_report on the target node, we
can use the information generated by the local configure script directly,
without searching.

The only preparation left to do is to create the subdirectory for the target.