Refactor: tools: implement crm_report collection command building in C
This implements building the commands needed to collect information locally and
from other nodes. This does not implement the execution of the commands.

Rather than copy a separate shell script (report.collector) to other nodes to
collect report information, the C version of crm_report adds a new --collect
option that the initiating node can pass to the other nodes, with a
colon-separated list of values needed for collection.

The .env file is no longer needed, and is not created.

Also, crm_report "fuzzes" the report times, collecting from 10 seconds before
the specified start time to 10 seconds after the specified end time. The
shell script version would fuzz the times at the start of collect_data(). Now,
the times are fuzzed when used.