WIP Fix: scheduler: block 'first' if 'then' is blocked restart
WIP all wrong.

For "start A then start B", if A is stopped (needing and able to start),
and B is started but with target-role=Stopped, then we won't schedule a B start
and thus the fix here will be irrelevant. But we still want to block the
A start or we'll lose the need to stop B if target-role is reset.