WIP Test: cts-scheduler: update expected outputs for ordering fix
** WIP before-unmanaged-started and before-unmanaged-stopped were not fixed,
and cloned_start_two should restart FAKECLONE2 on c7auto4. **

Now that we're adding an extra ordering for mandatory symmetric orderings,
those are showing up in the dot files, and the first actions as new
dependencies in the exp files. This is expected.

Certain tests have improved. In cases where the 'then' resource has to stop and
the 'first' resource has to start, the stop is now properly ordered before the
start when 'first' doesn't have to stop. These include bug-n-387749,
bundle-order-partial-start-2, cloned_start_one, order-mandatory,
order-required, order-wrong-kind, promoted-pseudo, and reload-becomes-restart.

bundle-order-stop-clone is odd, but it already was. A start pseudo-action was
scheduled for the bundle, and start and running pseudo-actions were scheduled
for the bundle clone, even though the bundled resource is stopping, not
starting. Now, the bundle pseudo-start is no longer scheduled, and the order of
the bundle clone pseudo-actions has changed.

Similarly, unrunnable-2 loses a bunch of dashed dependencies (indicating not in
the transition graph) in its dot file, with no other changes elsewhere. The
graph has only a blocked start. I'm not sure why these changed, but it seems
reasonable.