Refactor: pacemaker-attrd: track node CIB ID rather than cluster ID
Previously, attribute_value_t had a nodeid member to track the cluster
ID of the node that the value is for. However the only reason we need
it is to be able to get the node's XML ID in the CIB, for writing out
the value. Now, rename it to node_cib_id and track the XML ID directly.

In practice, there is no real change, since the CIB XML ID of Corosync
nodes is simply their cluster ID as a string. This allows us to keep
the same XML attribute and value in peer messages for backward
compatibility.

However this does mean that if we ever support node XML IDs that are
*not* the string equivalent of their cluster IDs, rolling upgrades will
be possible only from the version with this commit and later.