Refactor: scheduler: add pcmk__set_scheduler_defaults()
We currently have pe_reset_working_data(), cleanup_calculations(), and
set_working_set_defaults() for partially freeing and restoring defaults
in scheduler data.

The distinction is confusing, and this is the first step to
rationalizing it: the new function sets nonzero default values in
scheduler data, without freeing or nulling/zeroing anything, so it can
be used in both existing and replacement functions.